+++ Runtime/ModularAvatarWorldScaleObject.cs
@@ -1,0 +1,12 @@
+﻿using UnityEngine;
+
+namespace nadena.dev.modular_avatar.core
+{
+    [AddComponentMenu("Modular Avatar/MA World Scale Object")]
+    [DisallowMultipleComponent]
+    [HelpURL("https://modular-avatar.nadena.dev/docs/reference/world-scale-object?lang=auto")]
+    public class ModularAvatarWorldScaleObject : AvatarTagComponent
+    {
+        // no configuration
+    }
+}